Implementacja roli pracownika sortowni, do napisania kontroler, akcje oraz widoki

+++ Dispatch_system/Models/Employee.cs
@@ -14,7 +14,7 @@ namespace Dispatch_system.Models
         public int? PersonId { get; set; }
 
         [Required]
-        public int BranchId { get; set; }
+        public short BranchId { get; set; }
 
         [Required]
         public bool IsCourier { get; set; }
